feat(organizations): delete organization in org list

+++ next-tavla/app/(admin)/components/Delete/index.tsx
@@ -13,18 +13,39 @@ import { getFormFeedbackForField } from 'app/(admin)/utils'
 import { FormError } from '../FormError'
 import { useSearchParamsModal } from 'app/(admin)/hooks/useSearchParamsModal'
 import { deleteOrganization } from './actions'
+import { Tooltip } from '@entur/tooltip'
 
-function Delete({ organization }: { organization: TOrganization }) {
+function DeleteButton({ showText }: { showText?: boolean }) {
+    if (showText)
+        return (
+            <IconButton as={Link} href="?delete" className="g-2">
+                <DeleteIcon />
+                Slett
+            </IconButton>
+        )
+    return (
+        <Tooltip content="Slett organisasjon" placement="bottom">
+            <IconButton as={Link} href="?delete" className="g-2">
+                <DeleteIcon />
+            </IconButton>
+        </Tooltip>
+    )
+}
+
+function Delete({
+    organization,
+    showText,
+}: {
+    organization: TOrganization
+    showText?: boolean
+}) {
     const [modalIsOpen, close] = useSearchParamsModal('delete')
 
     const [state, action] = useFormState(deleteOrganization, undefined)
 
     return (
         <>
-            <IconButton as={Link} href="?delete" className="g-2">
-                <DeleteIcon />
-                Slett
-            </IconButton>
+            <DeleteButton showText={showText} />
             <Modal
                 open={modalIsOpen}
                 size="small"
